Move LocalDatabaseModule to root level
