+++ ground/src/androidTest/java/com/google/android/ground/TestLocalDatabaseModule.kt
@@ -17,7 +17,6 @@ package com.google.android.ground
 
 import android.content.Context
 import androidx.room.Room
-import com.google.android.ground.persistence.local.LocalDatabaseModule
 import com.google.android.ground.persistence.local.room.LocalDatabase
 import dagger.Module
 import dagger.Provides
