+++ ground/src/test/java/com/google/android/ground/persistence/local/TestSharedPreferencesModule.kt
@@ -17,6 +17,7 @@ package com.google.android.ground.persistence.local
 
 import android.content.Context
 import android.content.SharedPreferences
+import com.google.android.ground.LocalDatabaseModule
 import dagger.Module
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
